requirements.txt
appdirs==1.4.4
asgiref==3.3.4
audioread==2.1.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator==5.0.7
Django==3.2
django-cors-headers==3.7.0
django-rest-framework-base64-fields==0.1.3
djangorestframework==3.12.4
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
gunicorn==20.1.0
idna==2.10
joblib==1.0.1
llvmlite==0.36.0
numba==0.53.1
numpy==1.19.5
packaging==20.9
playsound==1.2.2
pooch==1.3.0
pybase64==1.1.4
pycparser==2.20
pydub==0.25.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
urllib3==1.26.4
